Refactor model selection utilities
